# apispec.yaml
swagger: '2.0'
consumes:
  - application/json
  - application/xml
produces:
  - application/json
  - application/xml
schemes:
  - http
  - https
info:
  title: 'Cerci API.'
  version: 0.0.1
host: localhost:8080
...
paths: 
...
  /project/{id}: 
    get:
      tags:
        - project
      summary: Find project item
      description: ''
      operationId: findProjectItem
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          id:
           format: int64
           type: integer
      responses:
        "200":
          $ref: '#/responses/ProjectResponse'  
        '404':
          description: News item  not found

  
    delete:
      tags:
        - project
      summary: Delete project item
      description: ''
      operationId: deleteProjectId
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          id:
           format: int64
           type: integer  
      responses:
        "200":
          $ref: '#/responses/ProjectResponse'  
        '404':
          description: Project item  not found                
...
responses:
...
  ProjectResponse:
    description: ProjectResponse response struct
    headers:
      detail:
        type: string
      finish_date:
        format: date-time
        type: string
      id:
        format: int64
        type: integer
      project_images:
        items:
          items:
            format: uint8
            type: integer
          type: array
        type: array
      project_name:
        type: string
      start_date:
        format: date-time
        type: string